split cuda_stream.h
